+++ packages/core/src/cookie.ts
@@ -277,7 +277,7 @@ export const createCookieStore = (
             ),
         },
         accessToken: {
-            name: `${securePrefix}${prefix}.${overrides?.accessToken?.name ?? "access_token"}`,
+            name: `${hostPrefix}${prefix}.${overrides?.accessToken?.name ?? "access_token"}`,
             attributes: defineSecureCookieOptions(
                 useSecure,
                 {